+++ src/formatter/terminal.rs
@@ -1,51 +1,63 @@
 #![allow(unused_must_use)]
 
 use super::Formatter;
-use crate::{constants::HIGHLIGHT_NAMES, FormatterOption, Options};
+use crate::languages::Language;
+use crate::{constants::HIGHLIGHT_NAMES, themes::Theme};
 use std::cell::RefCell;
 use std::io::Write;
 use termcolor::{ColorSpec, WriteColor};
-use tree_sitter_highlight::{Error, HighlightEvent};
+use tree_sitter_highlight::{HighlightEvent, Highlighter};
 
+#[derive(Clone, Debug)]
 pub struct Terminal<'a> {
-    options: Options<'a>,
     buffer: RefCell<termcolor::Buffer>,
+    source: &'a str,
+    lang: Language,
+    theme: Option<&'a Theme>,
 }
 
 impl<'a> Terminal<'a> {
-    pub fn new(options: Options<'a>) -> Self {
+    pub fn new(source: &'a str, lang: Language, theme: Option<&'a Theme>) -> Self {
         Self {
-            options,
             buffer: RefCell::new(termcolor::Buffer::ansi()),
+            source,
+            lang,
+            theme,
+        }
+    }
+}
+
+impl Default for Terminal<'_> {
+    fn default() -> Self {
+        Self {
+            buffer: RefCell::new(termcolor::Buffer::ansi()),
+            source: "",
+            lang: Language::PlainText,
+            theme: None,
         }
     }
 }
 
 impl Formatter for Terminal<'_> {
-    fn write<W>(
-        &self,
-        _writer: &mut W,
-        source: &str,
-        events: impl Iterator<Item = Result<HighlightEvent, Error>>,
-    ) where
-        W: std::fmt::Write,
-    {
-        // FIXME: implement italic
-        let _italic = if let FormatterOption::Terminal { italic } = &self.options.formatter {
-            *italic
-        } else {
-            false
-        };
+    fn highlights(&self) -> String {
+        let mut highlighter = Highlighter::new();
+        let events = highlighter
+            .highlight(
+                self.lang.config(),
+                self.source.as_bytes(),
+                None,
+                |injected| Some(Language::guess(injected, "").config()),
+            )
+            .expect("failed to generate highlight events");
 
         for event in events {
-            let event = event.expect("todo");
+            let event = event.expect("failed to get highlight event");
 
             match event {
                 HighlightEvent::HighlightStart(idx) => {
                     let scope = HIGHLIGHT_NAMES[idx.0];
 
                     let hex: &str = self
-                        .options
                         .theme
                         .as_ref()
                         .and_then(|theme| theme.get_style(scope))
@@ -63,7 +75,10 @@ impl Formatter for Terminal<'_> {
                         .set_color(ColorSpec::new().set_fg(Some(termcolor::Color::Rgb(r, g, b))));
                 }
                 HighlightEvent::Source { start, end } => {
-                    let text = source.get(start..end).expect("failed to get source bounds");
+                    let text = self
+                        .source
+                        .get(start..end)
+                        .expect("failed to get source bounds");
                     self.buffer.borrow_mut().write_all(text.as_bytes());
                 }
                 HighlightEvent::HighlightEnd => {
@@ -71,25 +86,7 @@ impl Formatter for Terminal<'_> {
                 }
             }
         }
-    }
 
-    fn finish<W>(&self, writer: &mut W, _: &str)
-    where
-        W: std::fmt::Write,
-    {
-        let output = String::from_utf8(self.buffer.borrow_mut().clone().into_inner()).unwrap();
-        let _ = writer.write_str(output.as_str());
-    }
-}
-
-impl Default for Terminal<'_> {
-    fn default() -> Self {
-        Self {
-            options: Options {
-                formatter: FormatterOption::Terminal { italic: false },
-                ..Options::default()
-            },
-            buffer: RefCell::new(termcolor::Buffer::ansi()),
-        }
+        String::from_utf8(self.buffer.borrow_mut().clone().into_inner()).unwrap()
     }
 }